Added interactions to the window
Still doesn't work exactly as intended

+++ project4a/project4a/ResultViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class ResultViewController: UIViewController {
     @IBOutlet weak var timeLabel: UILabel!
-    var time: Int?
+    var time: Double?
     override func viewDidLoad() {
         
     }
